change flash message setters to only log errors, not return them
In most cases, there isn't much you can do if flash message getting or
setting fails, beyond logging it. And certainly not much the end-user
can do. So simply using them as "try to set a flash message" makes our
code much simpler, and is closer to what we are actually doing anyway.