Update HTTP flow for errors on New User
Previously, if there was an error AFTER the user was created, it would
be rolled back by our standard HTTP sub-handler wrapper. This isn't what
I want, so we Commit right after the user is created (if it is).

I also found that the errors returned by the email failing was using the
sub-handler wrapper, whereas I think it would make more sense to
redirect them to the new user, but show them an error flash message.
It's a partial failure after all!