BUGFIX: "Ballot box with checkmark" emoji doesn't work on Linux
For whatever reason, the latest Ubuntu doesn't have a glyph for the
UTF-8 emoji I was using for indicating a user was IN a group. The empty
ballot box for indicating NOT IN a group does work. Looking over the
options, the "ballot box with x" does work in Linux and Windows. So I
simply made that change to deal with this "bug".